refact: moved Login handlers to context
