+++ src/contexts/authContext.js
@@ -1,5 +1,6 @@
 import { createContext, useContext, useState } from "react";
 import { useNavigate } from "react-router-dom";
+import { loginService } from "services";
 
 const AuthContext = createContext();
 
@@ -12,6 +13,35 @@ const AuthProvider = ({ children }) => {
     JSON.parse(localStorage.getItem("isAuth")) || false
   );
 
+  const [login, setLogin] = useState({
+    input: {},
+    error: "",
+    hide: { pwd: true },
+  });
+
+  const loginInputHandler = (e) => {
+    const { name, value } = e.target;
+    setLogin({ ...login, input: { ...login.input, [name]: value } });
+  };
+
+  const loginHandler = async (e) => {
+    e.preventDefault();
+
+    try {
+      const { data } = await loginService(login.input);
+      localStorage.setItem("isAuth", true);
+      localStorage.setItem("token", data.encodedToken);
+      setToken(data.encodedToken);
+
+      setLogin({ ...login, input: { email: "", password: "" } });
+      setIsAuth(true);
+
+      navigate("/home");
+    } catch (err) {
+      setLogin({ ...login, error: err.response.data.errors[0] });
+    }
+  };
+
   const logoutHandler = () => {
     localStorage.removeItem("token");
     localStorage.setItem("isAuth", false);
@@ -21,7 +51,18 @@ const AuthProvider = ({ children }) => {
 
   return (
     <AuthContext.Provider
-      value={{ isAuth, setIsAuth, token, setToken, navigate, logoutHandler }}
+      value={{
+        isAuth,
+        setIsAuth,
+        token,
+        setToken,
+        navigate,
+        login,
+        setLogin,
+        loginInputHandler,
+        loginHandler,
+        logoutHandler,
+      }}
     >
       {children}
     </AuthContext.Provider>
